Part 2/?

+++ source/features/quick-fork-deletion.tsx
@@ -3,39 +3,75 @@ import delay from 'delay';
 import React from 'dom-chef';
 import select from 'select-dom';
 import delegate from 'delegate-it';
+import elementReady from 'element-ready';
 import * as pageDetect from 'github-url-detection';
 
 import features from '.';
+import * as api from '../github-helpers/api';
 import {getRepo} from '../github-helpers';
 import pluralize from '../helpers/pluralize';
-import * as api from '../github-helpers/api';
-import elementReady from 'element-ready';
+import addNotice from '../github-widgets/notice-bar';
+import oneEvent from '../helpers/one-event';
 
-async function handleCancelClick(event: delegate.Event): Promise<void> {
+async function handleClick(event: delegate.Event): Promise<void> {
 	const button = event.delegateTarget;
 	event.preventDefault();
-	button.classList.add('btn-sm');
-	button.textContent = 'Delete fork';
+	if (!button.classList.toggle('btn-sm')) { // Use as state check
+		void start();
+	}
 }
 
-async function handleFirstClick(event: delegate.Event): Promise<void> {
-	const button = event.delegateTarget;
-	button.classList.remove('btn-sm');
-	event.preventDefault();
+async function buttonTimeout(button: HTMLElement): Promise<boolean> {
+	const abortController = new AbortController();
+	abortController.signal.addEventListener('abort', () => {
+		button.textContent = 'Delete fork';
+	});
+
+	// Watch for cancellations
+	button.addEventListener('click', () => abortController.abort(), {once: true});
 
 	let secondsLeft = 5;
 	do {
-		button.textContent = `Deleting fork in ${pluralize(secondsLeft, '$$ second')}`;
+		if (abortController.signal.aborted) {
+			return false;
+		}
+
+		button.textContent = `Deleting fork in ${pluralize(secondsLeft, '$$ second')}. Cancel?`;
 		await delay(1000); // eslint-disable-line no-await-in-loop
 	} while (--secondsLeft);
 
+	return true;
+}
+
+async function start(): Promise<void> {
+	const button = select<HTMLButtonElement>('.rgh-quick-fork-deletion')!;
+	if (!await buttonTimeout(button)) {
+		return;
+	}
+
+	button.disabled = true;
 	button.textContent = 'Deleting fork…';
 
-	alert('YO!');
-	console.log(await api.v3('', {
-		method: 'DELETE'
-	}));
-	alert('done');
+	const {nameWithOwner} = getRepo()!;
+	try {
+		await api.v3('/repos/' + nameWithOwner, {
+			method: 'DELETE'
+		});
+		addNotice(`Repository ${nameWithOwner} deleted`, {showCloseButton: false});
+		select('.application-main')!.remove();
+	} catch (error: unknown) {
+		button.remove();
+		addNotice(
+			<>
+				<p>Could not delete repository. Make sure <a href="https://github.com/settings/tokens">your token</a> has the <code>delete_repo</code> scope.</p>
+				<p>Full error: {(error as any).response?.message ?? (error as any).message}</p>
+			</>, {
+				showCloseButton: false
+			}
+		);
+
+		throw error;
+	}
 }
 
 async function init(): Promise<void> {
@@ -55,8 +91,7 @@ async function init(): Promise<void> {
 		</li>
 	);
 
-	delegate(document, '.rgh-quick-fork-deletion.btn-sm', 'click', handleFirstClick);
-	delegate(document, '.rgh-quick-fork-deletion', 'click', handleCancelClick);
+	delegate(document, '.rgh-quick-fork-deletion', 'click', handleClick);
 }
 
 void features.add(__filebasename, {
